rework dataviewOptionDelete

+++ core/block/editor/dataview/dataview.go
@@ -3,6 +3,7 @@ package dataview
 import (
 	"context"
 	"fmt"
+	"github.com/anytypeio/go-anytype-middleware/util/slice"
 	"strings"
 	"sync"
 
@@ -29,6 +30,7 @@ import (
 const defaultLimit = 50
 
 var log = logging.Logger("anytype-mw-editor")
+var ErrMultiupdateWasNotAllowed = fmt.Errorf("multiupdate was not allowed")
 
 type Dataview interface {
 	GetAggregatedRelations(blockId string) ([]*model.Relation, error)
@@ -43,7 +45,7 @@ type Dataview interface {
 	UpdateRelation(ctx *state.Context, blockId string, relationKey string, relation model.Relation, showEvent bool) error
 	AddRelationOption(ctx *state.Context, blockId string, recordId string, relationKey string, option model.RelationOption, showEvent bool) (*model.RelationOption, error)
 	UpdateRelationOption(ctx *state.Context, blockId string, recordId string, relationKey string, option model.RelationOption, showEvent bool) error
-	DeleteRelationOption(ctx *state.Context, blockId string, recordId string, relationKey string, optionId string, showEvent bool) error
+	DeleteRelationOption(ctx *state.Context, allowMultiupdate bool, blockId string, recordId string, relationKey string, optionId string, showEvent bool) error
 	FillAggregatedOptions(ctx *state.Context) error
 
 	CreateRecord(ctx *state.Context, blockId string, rec model.ObjectDetails, templateId string) (*model.ObjectDetails, error)
@@ -283,7 +285,7 @@ func (d *dataviewCollectionImpl) UpdateRelationOption(ctx *state.Context, blockI
 	return db.Update(recordId, []*model.Relation{rel}, database.Record{})
 }
 
-func (d *dataviewCollectionImpl) DeleteRelationOption(ctx *state.Context, blockId, recordId string, relationKey string, optionId string, showEvent bool) error {
+func (d *dataviewCollectionImpl) DeleteRelationOption(ctx *state.Context, allowMultiupdate bool, blockId, recordId string, relationKey string, optionId string, showEvent bool) error {
 	s := d.NewStateCtx(ctx)
 	tb, err := getDataviewBlock(s, blockId)
 	if err != nil {
@@ -303,14 +305,23 @@ func (d *dataviewCollectionImpl) DeleteRelationOption(ctx *state.Context, blockI
 		return fmt.Errorf("option id is empty")
 	}
 
-	err = tb.DeleteRelationOption(relationKey, optionId)
+	objIds, err := db.AggregateObjectIdsForOptionAndRelation(relationKey, optionId)
 	if err != nil {
 		return err
 	}
-	if showEvent {
-		err = d.Apply(s)
-	} else {
-		err = d.Apply(s, smartblock.NoEvent)
+
+	if len(objIds) > 1 && !allowMultiupdate {
+		return ErrMultiupdateWasNotAllowed
+	}
+
+	if slice.FindPos(objIds, recordId) == -1 {
+		// just in case we have some indexing lag
+		objIds = append(objIds, recordId)
+	}
+
+	err = tb.DeleteRelationOption(relationKey, optionId)
+	if err != nil {
+		return err
 	}
 
 	rel := pbtypes.GetRelation(tb.Model().GetDataview().Relations, relationKey)
@@ -318,7 +329,26 @@ func (d *dataviewCollectionImpl) DeleteRelationOption(ctx *state.Context, blockI
 		return fmt.Errorf("relation not found in dataview")
 	}
 
-	return db.Update(recordId, []*model.Relation{rel}, database.Record{})
+	for _, objId := range objIds {
+		err = db.Update(objId, []*model.Relation{rel}, database.Record{})
+		if err != nil {
+			if objId != recordId {
+				// not sure it is a right approach here, but we may face some ACL problems later otherwise
+				log.Errorf("DeleteRelationOption failed to multiupdate %s: %s", objId, err.Error())
+			} else {
+				return err
+			}
+		}
+		log.Debugf("DeleteRelationOption updated %s", objId)
+	}
+
+	if showEvent {
+		err = d.Apply(s)
+	} else {
+		err = d.Apply(s, smartblock.NoEvent)
+	}
+
+	return nil
 }
 
 func (d *dataviewCollectionImpl) GetAggregatedRelations(blockId string) ([]*model.Relation, error) {
